Raise ImportErrors when searching for tests in file patterns
test finder used to except ImportErrors when searching for test modules
in file patterns as a quick and easy way to silence errors due to
finding correct test files. However this causes a bug that if a test
file is found and is being imported, and the test file is actually
raising an ImportError (from the test or from one of the modules that
the test is importing), the error would not cause the test to fail as it
was being caught on the test runner, mistaking it by an error to find a
valid test file.

The fix for this is, first to try importing modules from real packages,
which now requires any directory containing test files to be a package,
that can be safely imported (so if an ImportError is raised from the
test file being imported, it really means it's an ImportError).
Since this breaks backward compatibility (not sure if anyone used
none-package subdirectories for their test files), an option will be
provided to except ImportErrors (CLI option will be added later),
to help broken tests.
To help this additional option avoid catching real ImportErrors, the
name of the module failed to be importd is checked agains the test file.
This way we try to (some level) avoid catching all ImportErrors.