Add except-import-errors CLI arg to test command
Had added the except_import_errors to catch ImportErrors raised while trying
to import test modules, to keep a backward compabitibile behavior.
So if anyone is relying on the fact that ImportErrors are caught when
importing test modules, they can use this option.

Normally this is not a desired behavior though.